requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
motor
python-dotenv
langdetect
sentencepiece
